# docs/openapi.yml
openapi: "3.0.3"
info:
  title: Bienenhaus Propiedades API
  description: |
    API RESTful del panel de administración de Bienenhaus Propiedades.
    Inmobiliaria boutique con CRM, tasaciones ACM, portales y redes sociales.
  version: "1.0.0"
  contact:
    name: Bienenhaus Propiedades
    url: https://bienenhaus.com.ar
servers:
  - url: https://bienenhaus.com.ar/api
    description: Producción
  - url: http://localhost:5000/api
    description: Desarrollo
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    Error:
      type: object
      properties:
        ok: { type: boolean, enum: [false] }
        error: { type: string }
    Property:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        type: { type: string, enum: [casa, departamento, terreno, local, galpon] }
        location: { type: string }
        price: { type: number }
        beds: { type: integer }
        baths: { type: integer }
        sqm: { type: number }
        sqm_total: { type: number }
        parkings: { type: integer }
        status: { type: string, enum: [disponible, reservado, vendido] }
        featured: { type: boolean }
        desc: { type: string }
        images: { type: array, items: { type: string } }
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
        created_at: { type: string, format: date-time }
        video_url: { type: string, nullable: true }
...
paths:
  /properties:
    get:
      summary: Lista propiedades
      security: [{ SessionAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: per_page
          schema: { type: integer, default: 20 }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string }
      responses:
        "200":
          description: Lista paginada de propiedades
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      items: { type: array, items: { $ref: "#/components/schemas/Property" } }
                      total: { type: integer }
                      pages: { type: integer }
                      page: { type: integer }
        "401":
          description: No autenticado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      summary: Crea una propiedad
      security: [{ SessionAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Property" }
      responses:
        "201":
          description: Propiedad creada
        "400":
          description: Datos inválidos
  "/properties/{id}":
    get:
      summary: Obtiene una propiedad
      security: [{ SessionAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Propiedad
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data: { $ref: "#/components/schemas/Property" }
        "404":
          description: No encontrada
    put:
      summary: Actualiza una propiedad
      security: [{ SessionAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Actualizada
    delete:
      summary: Elimina una propiedad
      security: [{ SessionAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Eliminada
...